requirements.txt
streamlit==1.39.0 
numpy==1.24.1 
polars==1.13.0 
matplotlib==3.6.3 
japanize-matplotlib==1.1.3 
streamlit-shadcn-ui==0.1.18 
PyYAML==6.0.2 
streamlit-authenticator==0.4.1